:recycle: 타입 컴포넌트 파일로 이동 및 타입파일 제거
