--- src/components/IndexTitle/IndexTitle.type.ts
@@ -1,8 +1,0 @@
-export interface IIndexTitleProps {
-  title: string;
-  total?: number;
-  clickHandler?: (idx: number) => void;
-  selectedIndex?: number;
-  onScreenCount?: number;
-  withoutIndex?: true;
-}
